+++ src/main/java/org/sosly/villageworks/profession/AbstractProfession.java
@@ -1,0 +1,22 @@
+package org.sosly.villageworks.profession;
+
+import net.minecraft.resources.ResourceLocation;
+import org.sosly.villageworks.api.profession.IProfession;
+
+public abstract class AbstractProfession implements IProfession {
+    private final ResourceLocation id;
+
+    protected AbstractProfession(ResourceLocation id) {
+        this.id = id;
+    }
+
+    @Override
+    public ResourceLocation getID() {
+        return id;
+    }
+
+    @Override
+    public String getTranslationKey() {
+        return "profession." + getID().toString().replace(":", ".");
+    }
+}